requirements.txt
absl-py==2.3.1
astunparse==1.6.3
attrs==25.4.0
beautifulsoup4==4.14.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
dlib-bin==19.24.6
face-recognition==1.3.0
face_recognition_models==0.3.0
filelock==3.20.0
Flask==3.1.2
flatbuffers==25.9.23
fonttools==4.60.1
grpcio==1.75.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
libclang==18.1.1
MarkupSafe==3.0.3
matplotlib==3.10.7
mediapipe==0.10.8
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.10.0.84
packaging==25.0
pandas==2.3.3
pillow==11.3.0
pycparser==2.23
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
scipy==1.15.3
sentencepiece==0.2.1
six==1.17.0
sounddevice==0.5.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==1.17.3
